requirements.txt
Django>=4.2
pymongo>=4.6
python-dotenv>=1.0.0
Pillow>=10.4.0
bcrypt>=4.1.0
dnspython>=2.4.0
certifi>=2023.7.22
